+++ src/RestQuery.js
@@ -377,7 +377,11 @@ RestQuery.prototype.handleInclude = function() {
       this.include = this.include.slice(1);
       return this.handleInclude();
     });
+  } else if (this.include.length > 0) {
+    this.include = this.include.slice(1);
+    return this.handleInclude();
   }
+  
   return pathResponse;
 };
 
@@ -415,6 +419,11 @@ function includePath(config, auth, response, path) {
     for (var obj of includeResponse.results) {
       obj.__type = 'Object';
       obj.className = className;
+
+      if(className == "_User"){
+        delete obj.sessionToken;
+      }
+
       replace[obj.objectId] = obj;
     }
     var resp = {