+++ src/schemas.js
@@ -7,36 +7,27 @@ var express = require('express'),
 
 var router = new PromiseRouter();
 
-function mongoFieldTypeToSchemaAPIType(type) {
-  if (type[0] === '*') {
-    return {
-      type: 'Pointer',
-      targetClass: type.slice(1),
-    };
-  }
-  if (type.startsWith('relation<')) {
-    return {
-      type: 'Relation',
-      targetClass: type.slice('relation<'.length, type.length - 1),
-    };
-  }
-  switch (type) {
-    case 'number':   return {type: 'Number'};
-    case 'string':   return {type: 'String'};
-    case 'boolean':  return {type: 'Boolean'};
-    case 'date':     return {type: 'Date'};
-    case 'map':
-    case 'object':   return {type: 'Object'};
-    case 'array':    return {type: 'Array'};
-    case 'geopoint': return {type: 'GeoPoint'};
-    case 'file':     return {type: 'File'};
-  }
+function masterKeyRequiredResponse() {
+  return Promise.resolve({
+    status: 401,
+    response: {error: 'master key not specified'},
+  })
+}
+
+function classNameMismatchResponse(bodyClass, pathClass) {
+  return Promise.resolve({
+    status: 400,
+    response: {
+      code: Parse.Error.INVALID_CLASS_NAME,
+      error: 'class name mismatch between ' + bodyClass + ' and ' + pathClass,
+    }
+  });
 }
 
 function mongoSchemaAPIResponseFields(schema) {
   var fieldNames = Object.keys(schema).filter(key => key !== '_id' && key !== '_metadata');
   var response = fieldNames.reduce((obj, fieldName) => {
-    obj[fieldName] = mongoFieldTypeToSchemaAPIType(schema[fieldName])
+    obj[fieldName] = Schema.mongoFieldTypeToSchemaAPIType(schema[fieldName])
     return obj;
   }, {});
   response.ACL = {type: 'ACL'};
@@ -55,10 +46,7 @@ function mongoSchemaToSchemaAPIResponse(schema) {
 
 function getAllSchemas(req) {
   if (!req.auth.isMaster) {
-    return Promise.resolve({
-      status: 401,
-      response: {error: 'master key not specified'},
-    });
+    return masterKeyRequiredResponse();
   }
   return req.config.database.collection('_SCHEMA')
   .then(coll => coll.find({}).toArray())
@@ -69,10 +57,7 @@ function getAllSchemas(req) {
 
 function getOneSchema(req) {
   if (!req.auth.isMaster) {
-    return Promise.resolve({
-      status: 401,
-      response: {error: 'unauthorized'},
-    });
+    return masterKeyRequiredResponse();
   }
   return req.config.database.collection('_SCHEMA')
   .then(coll => coll.findOne({'_id': req.params.className}))
@@ -88,20 +73,11 @@ function getOneSchema(req) {
 
 function createSchema(req) {
   if (!req.auth.isMaster) {
-    return Promise.resolve({
-      status: 401,
-      response: {error: 'master key not specified'},
-    });
+    return masterKeyRequiredResponse();
   }
   if (req.params.className && req.body.className) {
     if (req.params.className != req.body.className) {
-      return Promise.resolve({
-        status: 400,
-        response: {
-          code: Parse.Error.INVALID_CLASS_NAME,
-          error: 'class name mismatch between ' + req.body.className + ' and ' + req.params.className,
-        },
-      });
+      return classNameMismatchResponse(req.body.className, req.params.className);
     }
   }
   var className = req.params.className || req.body.className;
@@ -123,9 +99,94 @@ function createSchema(req) {
   }));
 }
 
+function modifySchema(req) {
+  if (!req.auth.isMaster) {
+    return masterKeyRequiredResponse();
+  }
+
+  if (req.body.className && req.body.className != req.params.className) {
+    return classNameMismatchResponse(req.body.className, req.params.className);
+  }
+
+  var submittedFields = req.body.fields || {};
+  var className = req.params.className;
+
+  return req.config.database.loadSchema()
+  .then(schema => {
+    if (!schema.data[className]) {
+      return Promise.resolve({
+        status: 400,
+        response: {
+          code: Parse.Error.INVALID_CLASS_NAME,
+          error: 'class ' + req.params.className + ' does not exist',
+        }
+      });
+    }
+    var existingFields = schema.data[className];
+
+    for (var submittedFieldName in submittedFields) {
+      if (existingFields[submittedFieldName] && submittedFields[submittedFieldName].__op !== 'Delete') {
+        return Promise.resolve({
+          status: 400,
+          response: {
+            code: 255,
+            error: 'field ' + submittedFieldName + ' exists, cannot update',
+          }
+        });
+      }
+
+      if (!existingFields[submittedFieldName] && submittedFields[submittedFieldName].__op === 'Delete') {
+        return Promise.resolve({
+          status: 400,
+          response: {
+            code: 255,
+            error: 'field ' + submittedFieldName + ' does not exist, cannot delete',
+          }
+        });
+      }
+    }
+
+    var newSchema = Schema.buildMergedSchemaObject(existingFields, submittedFields);
+    var mongoObject = Schema.mongoSchemaFromFieldsAndClassName(newSchema, className);
+    if (!mongoObject.result) {
+      return Promise.resolve({
+        status: 400,
+        response: mongoObject,
+      });
+    }
+
+    // Finally we have checked to make sure the request is valid and we can start deleting fields.
+    // Do all deletions first, then a single save to _SCHEMA collection to handle all additions.
+    var deletionPromises = []
+    Object.keys(submittedFields).forEach(submittedFieldName => {
+      if (submittedFields[submittedFieldName].__op === 'Delete') {
+        var promise = req.config.database.connect()
+        .then(() => schema.deleteField(
+          submittedFieldName,
+          className,
+          req.config.database.db,
+          req.config.database.collectionPrefix
+        ));
+        deletionPromises.push(promise);
+      }
+    });
+
+    return Promise.all(deletionPromises)
+    .then(() => new Promise((resolve, reject) => {
+      schema.collection.update({_id: className}, mongoObject.result, {w: 1}, (err, docs) => {
+        if (err) {
+          reject(err);
+        }
+        resolve({ response: mongoSchemaToSchemaAPIResponse(mongoObject.result)});
+      })
+    }));
+  });
+}
+
 router.route('GET', '/schemas', getAllSchemas);
 router.route('GET', '/schemas/:className', getOneSchema);
 router.route('POST', '/schemas', createSchema);
 router.route('POST', '/schemas/:className', createSchema);
+router.route('PUT', '/schemas/:className', modifySchema);
 
 module.exports = router;