+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 ## Parse Server Changelog
 
+### 2.1.0 (2/17/2016)
+
+* Feature: Support for additional OAuth providers
+* Feature: Ability to implement custom OAuth providers
+* Feature: Support for deleting Parse Files
+* Feature: Allow querying roles
+* Feature: Support for logs, extensible via Log Adapter
+* Feature: New Push Adapter for sending push notifications through OneSignal
+* Feature: Tighter default security for Users
+* Feature: Pass parameters to Cloud Code in query string
+* Feature: Disable anonymous users via configuration.
+* Experimental: Schemas API support for PUT operations
+* Fix: Prevent installation ID from being added to User
+* Fix: Becoming a user works properly with sessions
+* Fix: Including multiple object when some object are unavailable will get all the objects that are available
+* Fix: Invalid URL for Parse Files
+* Fix: Making a query without a limit now returns 100 results
+* Fix: Expose installation id in cloud code
+* Fix: Correct username for Anonymous users
+* Fix: Session token issue after fetching user
+* Fix: Issues during install process
+* Fix: Issue with Unity SDK sending _noBody
+
 ### 2.0.8 (2/11/2016)
 
 * Add: support for Android and iOS push notifications